refactor: colocate trace connectivity helper
